Moved all tables into `fuel-core-interfaces/src/db` (#788)
* Moved all tablets into `fuel-core-interfaces/src/db`.

* Return back `balances.rs`

* Make CI happy
